# Cache replicated archetypes for faster iteration