--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Cache replicated archetypes for faster iteration.
+
 ## [0.18.0] - 2023-12-19
 
 ### Changed