Gives nice speedup according to our benchmarks:
```
     Running benches/replication.rs (target/release/deps/replication-c5d7769abd7ff06d)
entities send           time:   [63.425 µs 63.526 µs 63.622 µs]
                        change: [-11.110% -10.897% -10.703%] (p = 0.00 < 0.05)
                        Performance has improved.

entities receive        time:   [135.24 µs 135.43 µs 135.67 µs]
                        change: [-2.9535% -2.3766% -1.8401%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 20 measurements (5.00%)
  1 (5.00%) high mild

entities update send    time:   [62.996 µs 63.392 µs 63.810 µs]
                        change: [-10.514% -8.9546% -7.0856%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 20 measurements (10.00%)
  1 (5.00%) high mild
  1 (5.00%) high severe

entities update receive time:   [67.951 µs 68.413 µs 68.944 µs]
                        change: [-1.9278% -0.9178% +0.0163%] (p = 0.08 > 0.05)
                        No change in performance detected.
Found 1 outliers among 20 measurements (5.00%)
  1 (5.00%) high severe
```

But I suspect that it will be a lot faster then the previous approach when there are many archetypes.