@Shatur 2023-12-20
`@NiseVoid` in Discord mentioned it.
It's similar to how `QueryState` works, but with access-info stripped out. I borrowed the code from it.